Add files via upload
A regression model elaborated to explain birthweights in the United States.
The model follows a standard process for data treatment. It starts with the exploration of the dataset, it continues with the imputation of missing values and it concludes with feature engineering and the choice of the best model.
This final step privileged the explanatory capacity instead of the reduction of gap between train and test model. The maximum level for the gap was set to 0.05.